feat(stn): use f64 instead of f32 for improved precision

+++ rustamer-base/src/search_state.rs
@@ -35,7 +35,7 @@ use super::utils::*;
 #[derive(Debug)]
 pub struct State {
     pub assignments: Vector<ExpressionNode>,
-    pub temporal_network: Option<DeltaSTN<u64, f32>>,
+    pub temporal_network: Option<DeltaSTN<u64, f64>>,
     pub todo: FxHashMap<Action, (usize, u32)>,
     pub active_conditions: HashMultiSet<Vec<ExpressionNode>>,
     pub g: f64,
